:tada: Explore splits by passing --explore <split> to edexplore
edexplore is not yet compatible with the dev branch as it expects
`dataset` as key in the config, wheres edflow expects `datasets/train`
and `datasets/validation`.
edexplore can now handle both cases and you can now specify which split
you want to explore by defineing the key `explore: <split>` in the
config either by storing the value in the corresponding yaml file or by
passing `--explore <split>` to edexplore.